(#118) Relax bundler requirement
Prior to this commit, rubycas-server depended on the bundler gem even
in environments in which the necessary dependencies are already met,
such as some production environments. This forces users to
unnecessarily install bundler, which isn't an option in many
restrictive production environments. With this commit, the bundler
dependency is relaxed so that a failure to load bundler doesn't result
in failure, but rather allows execution to continue under the
assumption that the necessary dependencies are already in place. This
commit should not change any behavior in environments where bundler is
available.