Add restricted_domain option to Google authenticator
Previous to this commit, the Google authenticator would authenticate
any Google account given proper credentials. This commit introduces
the restricted_domain Google authenticator configuration option in
order to restrict authentication only to those Google accounts for
which both proper credentials are given and the domain of the account
matches the restricted_domain option. If no restricted_domain option
is given, the Google authenticator will have the same behavior as
previous to this commit.