Add migration for previously added authenticator column in casserver_tgt
Previous to this commit, an up-to-date RubyCAS-Server would crash upon
encountering a session that was created before the "authenticator"
column had been added to the casserver_tgt table. This commit is not a
full fix for this problem, but it does prevent the problem from
occurring if the session is retrieved from the server rather than the
cookie.