Use a defined session key
Previous to this commit, RubyCAS server did not use a session key to
check the validity of the 'tgt' cookie. This made it it impossible to
invalidate all cookies containing a Ticket Granting Ticket.

This commit adds a **session_secret** key to the config.yml file.
Further, the config.ru now uses Rack::Session::Cookie to manage the
cookies, using the session_secret setting in the config file.